Refactor Room package structure and implement SharedPreferences and SQLite controllers for record management
