+++ app/src/main/java/com/SarayDani/sidi/data/room/EntidadRecord.kt
@@ -1,4 +1,4 @@
-package com.SarayDani.sidi.controller.room
+package com.SarayDani.sidi.data.room
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
